refactor: use-get-footnote-body로 파일 이름 변경 (#447)
