+++ src/components/mdx-components/footnote-tooltip.tsx
@@ -1,6 +1,6 @@
 import { CSSProperties, useEffect, useRef } from 'react';
 
-import useGetFootnoteBody from '@/hooks/useGetFootnoteBody';
+import useGetFootnoteBody from '@/hooks/use-get-footnote-body';
 
 interface Props {
   idx: string;
@@ -56,7 +56,7 @@ const FootnoteTooltip = ({ idx, show }: Props) => {
     <span
       aria-label="footnote tooltip"
       ref={ref}
-      className="invisible absolute left-0 z-30 w-max whitespace-nowrap rounded-md border border-gray-700 bg-white px-5 drop-shadow dark:border-gray-700 dark:border-gray-300 dark:bg-gray-900 sm:max-w-md md:max-w-xl lg:max-w-2xl xl:max-w-3xl 2xl:max-w-7xl"
+      className="invisible absolute left-0 z-30 w-max whitespace-nowrap rounded-md border border-gray-700 bg-white px-5 drop-shadow dark:border-gray-300 dark:border-gray-700 dark:bg-gray-900 sm:max-w-md md:max-w-xl lg:max-w-2xl xl:max-w-3xl 2xl:max-w-7xl"
       style={getStyle() as CSSProperties}
     >
       <span
